refactor: cart 관련 함수 분리
